+++ demo/physics_demo.py
@@ -16,7 +16,7 @@ rb.Game.show_fps = True
 main_scene = rb.Scene()  # Create our scene
 
 # Create our four walls
-top = rb.GameObject(pos=rb.Display.top_center + rb.Vector(0, -60)).add(
+top = rb.GameObject(pos=rb.Display.top_center + rb.Vector(0, 60)).add(
     rb.Rectangle(
         width=rb.Display.res.x + 175,
         height=rb.Display.res.y // 10,
@@ -24,7 +24,7 @@ top = rb.GameObject(pos=rb.Display.top_center + rb.Vector(0, -60)).add(
     )
 )
 
-bottom = rb.GameObject(pos=rb.Display.bottom_center + rb.Vector(0, 60)).add(
+bottom = rb.GameObject(pos=rb.Display.bottom_center + rb.Vector(0, -60)).add(
     rb.Rectangle(
         width=rb.Display.res.x + 175,
         height=rb.Display.res.y // 10,
@@ -56,18 +56,21 @@ for _ in range(num_obj // 2):
     main_scene.add(
         rb.wrap(
             [
-                rb.Circle(radius=rb.Display.res.x // num_obj, color=rb.Color.random_default()),
+                rb.Circle(
+                    radius=rb.Display.res.x // num_obj,
+                    color=rb.Color.random_default(),
+                ),
                 rb.RigidBody(
                     mass=0.1,
                     bounciness=0.99,
                     friction=0.2,
-                    gravity=(0, 80),
+                    gravity=(0, -80),
                     velocity=(randint(-100, 100), randint(-100, 100)),
                 ),
             ],
-            pos=(
+            pos=rb.Display.top_left + (
                 randint(int(rb.Display.res.x / 20), int(19 * rb.Display.res.x / 20)),
-                randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
+                -randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
             ),
         )
     )
@@ -75,18 +78,21 @@ for _ in range(num_obj // 2):
     main_scene.add(
         rb.wrap(
             [
-                rb.Polygon(rb.Vector.poly(randint(3, 9), rb.Display.res.x // num_obj), color=rb.Color.random_default()),
+                rb.Polygon(
+                    rb.Vector.poly(randint(3, 9), rb.Display.res.x // num_obj),
+                    color=rb.Color.random_default(),
+                ),
                 rb.RigidBody(
                     mass=0.1,
                     bounciness=0.99,
                     friction=0.2,
-                    gravity=(0, 80),
+                    gravity=(0, -80),
                     velocity=(randint(-100, 100), randint(-100, 100)),
                 ),
             ],
-            pos=(
+            pos=rb.Display.top_left + (
                 randint(int(rb.Display.res.x / 20), int(19 * rb.Display.res.x / 20)),
-                randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
+                -randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
             ),
         )
     )