+++ demo/platformer.py
@@ -74,13 +74,13 @@ player.add(rb.Rectangle(width=40, height=64, tag="player"))
 player.add(rb.Rectangle(
     width=34,
     height=2,
-    offset=rb.Vector(0, 32),
+    offset=rb.Vector(0, -32),
     trigger=True,
     on_collide=player_collide,
 ))
 
 # define the player rigidbody
-player_body = rb.RigidBody(gravity=rb.Vector(y=rb.Display.res.y * 1.5), pos_correction=1)
+player_body = rb.RigidBody(gravity=rb.Vector(y=rb.Display.res.y * -1.5), pos_correction=1)
 player.add(player_body)
 
 # Side boundary
@@ -97,8 +97,8 @@ ground.get(rb.Rectangle).bottom_left = rb.Display.bottom_left
 # create platforms
 platforms = [
     rb.GameObject(pos=rb.Vector(
-        200,
-        rb.Display.bottom - 140,
+        -760,
+        rb.Display.bottom + 140,
     )).add(rb.Rectangle(
         width=90,
         height=40,
@@ -106,8 +106,8 @@ platforms = [
         color=rb.Color.blue,
     )),
     rb.GameObject(pos=rb.Vector(
-        400,
-        rb.Display.bottom - 340,
+        -560,
+        rb.Display.bottom + 340,
     )).add(rb.Rectangle(
         width=150,
         height=40,
@@ -118,13 +118,13 @@ platforms = [
 
 # create obstacles
 obstacles = [
-    rb.GameObject(pos=rb.Vector(700)).add(rb.Rectangle(
+    rb.GameObject(pos=rb.Vector(-260)).add(rb.Rectangle(
         width=90,
         height=500,
         tag="ground",
         color=rb.Color.purple,
     )),
-    rb.GameObject(pos=rb.Vector(1200)).add(rb.Rectangle(
+    rb.GameObject(pos=rb.Vector(240)).add(rb.Rectangle(
         width=70,
         height=450,
         tag="ground",
@@ -133,13 +133,13 @@ obstacles = [
 ]
 
 for obstacle in obstacles:
-    obstacle.get(rb.Rectangle).bottom = rb.Display.bottom - 30
+    obstacle.get(rb.Rectangle).bottom = rb.Display.bottom + 30
 
 # create triggers
 triggers = [
     rb.GameObject(pos=rb.Vector(
-        950,
-        rb.Display.bottom - 45,
+        -10,
+        rb.Display.bottom + 45,
     )).add(rb.Rectangle(
         width=400,
         height=30,
@@ -159,7 +159,7 @@ portal_animation = rb.Animation(scale=rb.Vector(4, 4), fps=2)
 portal_animation.add_spritesheet("", all_portal_images, to_coord=all_portal_images.end)
 
 # create the end portal
-portal = rb.GameObject(pos=rb.Display.bottom_left + rb.Vector(level_size - 50, -100))
+portal = rb.GameObject(pos=rb.Display.bottom_left + rb.Vector(level_size - 50, 100))
 portal.add(portal_animation)
 
 portal.add(
@@ -182,7 +182,6 @@ def update():
 
     move_axis = rb.Input.controller_axis(rb.Input.controllers - 1, 0)
     centered = rb.Input.axis_centered(move_axis)
-
     # check for user directional input
     if rb.Input.key_pressed("a") or (move_axis < 0 and not centered):
         player_body.velocity.x = -300
@@ -210,7 +209,7 @@ def update():
 
     if rb.Input.key_pressed("r") or rb.Input.controller_button(rb.Input.controllers - 1, 6):
         player.pos = rb.Display.center_left + rb.Vector(50, 0)
-        player.get(rb.RigidBody).velocity = rb.Vector.zero()
+        player.get(rb.RigidBody).stop()
         grounded = False
 
     if rb.Input.key_pressed("space"):
@@ -221,7 +220,7 @@ def update():
 def fixed_update():
     # have the camera follow the player
     camera_ideal = rb.Math.clamp(
-        player.pos.x + rb.Display.res.x / 4, rb.Display.center.x, level_size - rb.Display.res.x / 2
+        player.pos.x + rb.Display.res.x / 4, rb.Display.center.x, level_size - rb.Display.res.x
     )
     rb.Game.camera.pos.x = rb.Math.lerp(rb.Game.camera.pos.x, camera_ideal, rb.Time.fixed_delta / 0.4)
 
@@ -236,7 +235,7 @@ def handle_keydown(event):
     global jumps, grounded
     if event["key"] == "w" and jumps > 0:
         grounded = False
-        player_body.velocity.y = -800
+        player_body.velocity.y = 800
         if jumps == 2:
             p_animation.set_state("jump", freeze=2)
         elif jumps == 1: