+++ demo/tetris.py
@@ -59,8 +59,11 @@ pos = V(1, 0)
 
 
 def draw_block(pos, color):
+    ctr = Display.top_left
+    ctr.x += (pos.x + 1) * block_size
+    ctr.y += (pos.y - 1) * block_size
     Draw.rect(
-        center=pos * block_size + 1 + block_size / 2,
+        center=ctr,
         width=block_size,
         height=block_size,
         border=None,
@@ -74,18 +77,18 @@ def draw_scene():
     for i in range(len(field)):
         for j in range(len(field[i])):
             if field[i][j]:
-                draw_block(V(j, i), Color.magenta)
+                draw_block(V(j, -i), Color.magenta)
             else:
-                draw_block(V(j, i), Color.black.lighter())
+                draw_block(V(j, -i), Color.black.lighter())
 
     for i in range(len(current)):
         for j in range(len(current[i])):
             if current[i][j]:
-                draw_block(pos + V(j, i), Color.red)
+                draw_block(pos + V(j, -i), Color.red)
 
 
 def moveDown():
-    pos.y += 1
+    pos.y -= 1
     incScore(1)